Emscripten
- cast size_t in for loop to int

+++ Minigin/GameObject.h
@@ -27,7 +27,7 @@ namespace dae
 		template<typename T>
 		void RemoveComponent()
 		{
-			for (int index{0}; index < m_components.size(); ++index)
+			for (int index{0}; index < (int)m_components.size(); ++index)
 			{
 				if (T* removeComp = dynamic_cast<T*>(m_components[index]))
 				{
